day2/pyproject.toml
[project]
name = "day2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "google-generativeai>=0.8.6",
    "openai>=2.24.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
